feat(telemetry): mur.routing outcome recording from FallbackLlmClient (best-effort, all intents)
- Event::Routing variant (telemetry_writer.rs) + event_to_notification arm
  emitting the new mur.routing method (mur-common/src/telemetry.rs const).
- FallbackLlmClient gains an optional telemetry sender + agent name
  (with_telemetry builder, additive-only — existing new()/new_routed() call
  sites and unit tests are unaffected). After the fallback/cascade loop
  resolves (Ok or final Err), it emits one Event::Routing via try_send
  (never awaited, never fails the turn).
- build_provider_runner (supervisor_runner.rs) takes an optional telemetry
  sender and wires it into the routed FallbackLlmClient; supervisor.rs
  passes Some(writer.sender()) at its one production call site.
- Split mur-agent-runtime/src/llm/fallback.rs (grew to 859 lines) into
  fallback/mod.rs (production, 467 lines) + fallback/tests.rs (390 lines)
  to satisfy the repo's 800-line file cap — pure code movement, no
  behavior change.
