+++ mur-agent-runtime/src/llm/fallback/tests.rs
@@ -1,0 +1,390 @@
+use super::super::BackgroundKind;
+use super::*;
+
+#[test]
+fn cooldown_marks_and_expires() {
+    let cm = CooldownMap::new();
+    let now = Instant::now();
+    assert!(!cm.is_cooling("m", now));
+    cm.mark("m", now + Duration::from_secs(60));
+    assert!(cm.is_cooling("m", now)); // within window
+    assert!(!cm.is_cooling("m", now + Duration::from_secs(61))); // after window
+    assert!(!cm.is_cooling("other", now));
+}
+
+#[test]
+fn backoff_grows_and_stays_in_bounds() {
+    let base = 500u64;
+    for attempt in 0..4u32 {
+        let d = backoff_delay(attempt, base).as_millis() as u64;
+        let floor = base * 2u64.pow(attempt);
+        assert!(
+            d >= floor && d < floor + base,
+            "attempt {attempt}: {d} not in [{floor}, {})",
+            floor + base
+        );
+    }
+}
+
+#[test]
+fn estimate_tokens_sums_text_over_rich_messages() {
+    let req = LlmRequest {
+        messages: vec![
+            RichMessage::Text {
+                role: "user".into(),
+                content: "a".repeat(40),
+            },
+            RichMessage::ImageText {
+                role: "user".into(),
+                media_type: "image/png".into(),
+                data: String::new(),
+                text: "b".repeat(40),
+            },
+        ],
+        temperature: None,
+        max_tokens: None,
+        tools: vec![],
+        ..Default::default()
+    };
+    assert_eq!(estimate_input_tokens(&req), 20); // 80 chars / 4
+}
+
+use super::super::StopReason;
+use std::sync::atomic::{AtomicUsize, Ordering};
+
+// LlmResponse has no Default (StopReason has no default variant), so build
+// one explicitly.
+fn mk_resp(text: &str) -> LlmResponse {
+    LlmResponse {
+        text: text.to_string(),
+        input_tokens: 0,
+        output_tokens: 0,
+        model: text.to_string(),
+        tool_calls: vec![],
+        stop_reason: StopReason::EndTurn,
+    }
+}
+
+// Mock client whose Nth generate() outcome is scripted.
+struct ScriptClient {
+    name: String,
+    outcomes: Vec<Result<(), LlmError>>,
+    idx: AtomicUsize,
+}
+#[async_trait]
+impl LlmClient for ScriptClient {
+    async fn generate(&self, _req: LlmRequest) -> Result<LlmResponse, LlmError> {
+        let i = self
+            .idx
+            .fetch_add(1, Ordering::SeqCst)
+            .min(self.outcomes.len() - 1);
+        match &self.outcomes[i] {
+            Ok(()) => Ok(mk_resp(&self.name)),
+            Err(e) => Err(e.clone()),
+        }
+    }
+    fn model_name(&self) -> &str {
+        &self.name
+    }
+}
+
+fn factory_for(scripts: HashMap<String, Vec<Result<(), LlmError>>>) -> ClientFactory {
+    Box::new(move |r: &str| {
+        let o = scripts.get(r).cloned().unwrap_or_else(|| vec![Ok(())]);
+        Ok(Arc::new(ScriptClient {
+            name: r.to_string(),
+            outcomes: o,
+            idx: AtomicUsize::new(0),
+        }) as Arc<dyn LlmClient>)
+    })
+}
+
+fn retry0() -> RetryConfig {
+    RetryConfig {
+        max_retries: 0,
+        backoff_base_ms: 1,
+        cooldown_secs: 60,
+    }
+}
+
+#[tokio::test]
+async fn advances_chain_on_retryable_then_succeeds() {
+    let mut s = HashMap::new();
+    s.insert("a".into(), vec![Err(LlmError::ServerError(500))]); // a fails (retryable)
+    s.insert("b".into(), vec![Ok(())]); // b succeeds
+    let fb = FallbackLlmClient::new(vec!["a".into(), "b".into()], factory_for(s), retry0());
+    let resp = fb.generate(LlmRequest::default()).await.unwrap();
+    assert_eq!(resp.text, "b"); // fell through to b
+}
+
+#[tokio::test]
+async fn fatal_error_does_not_advance() {
+    let mut s = HashMap::new();
+    s.insert("a".into(), vec![Err(LlmError::Http("401".into()))]); // fatal
+    s.insert("b".into(), vec![Ok(())]);
+    let fb = FallbackLlmClient::new(vec!["a".into(), "b".into()], factory_for(s), retry0());
+    let err = fb.generate(LlmRequest::default()).await.unwrap_err();
+    assert!(matches!(err, LlmError::Http(_))); // returned a's fatal error, never tried b
+}
+
+#[tokio::test]
+async fn exhaustion_returns_last_error() {
+    let mut s = HashMap::new();
+    s.insert("a".into(), vec![Err(LlmError::RateLimit)]);
+    s.insert("b".into(), vec![Err(LlmError::ServerError(503))]);
+    let fb = FallbackLlmClient::new(vec!["a".into(), "b".into()], factory_for(s), retry0());
+    let err = fb.generate(LlmRequest::default()).await.unwrap_err();
+    assert!(matches!(err, LlmError::ServerError(503))); // last candidate's error
+}
+
+#[test]
+fn candidates_pin_overrides_everything() {
+    // A pinned ref returns exactly [pinned] regardless of source.
+    let fb = FallbackLlmClient::new(
+        vec!["a".into(), "b".into()],
+        factory_for(Default::default()),
+        retry0(),
+    );
+    let mut req = LlmRequest::default();
+    req.pin_model_ref = Some("frontier".into());
+    assert_eq!(fb.candidates_for(&req), vec!["frontier".to_string()]);
+}
+
+#[test]
+fn candidates_smart_background_prepends_cheap() {
+    use mur_common::agent::AgentProfile;
+    use mur_common::config::{ModelSwitchConfig, SmartConfig};
+    let mut cfg = ModelSwitchConfig::default();
+    cfg.default = Some("primary".into());
+    cfg.fallback_chain = vec!["primary".into(), "mid".into()];
+    cfg.smart = SmartConfig {
+        enabled: true,
+        cheap: Some("cheap".into()),
+        max_escalations: 1,
+    };
+    let fb = FallbackLlmClient::new_routed(
+        AgentProfile::default_for_tests(),
+        cfg,
+        factory_for(Default::default()),
+        retry0(),
+    );
+    // Background + smart → cheap first, then phase-1 candidates, deduped.
+    let mut bg = LlmRequest::default();
+    bg.intent = RequestIntent::Background(BackgroundKind::Scheduled);
+    assert_eq!(
+        fb.candidates_for(&bg),
+        vec!["cheap".to_string(), "primary".into(), "mid".into()]
+    );
+    // Interactive → unchanged (no cheap prepend).
+    let inter = LlmRequest::default();
+    assert_eq!(
+        fb.candidates_for(&inter),
+        vec!["primary".to_string(), "mid".into()]
+    );
+}
+
+#[tokio::test]
+async fn routed_generate_picks_frontier_for_large_request() {
+    use mur_common::agent::AgentProfile;
+    use mur_common::config::{ModelSwitchConfig, RoutingConfig};
+    let mut cfg = ModelSwitchConfig::default();
+    cfg.routing = RoutingConfig {
+        enabled: true,
+        cheap: Some("cheap".into()),
+        frontier: Some("frontier".into()),
+        threshold_input_tokens: Some(5),
+        smart: None,
+    };
+    let mut scripts = std::collections::HashMap::new();
+    scripts.insert("frontier".to_string(), vec![Ok(())]);
+    scripts.insert("cheap".to_string(), vec![Ok(())]);
+    let fb = FallbackLlmClient::new_routed(
+        AgentProfile::default_for_tests(),
+        cfg,
+        factory_for(scripts),
+        retry0(),
+    );
+    // A big request (> threshold=5 tokens) routes to frontier.
+    let big = LlmRequest {
+        messages: vec![RichMessage::Text {
+            role: "user".into(),
+            content: "x".repeat(400),
+        }],
+        temperature: None,
+        max_tokens: None,
+        tools: vec![],
+        ..Default::default()
+    };
+    assert_eq!(fb.generate(big).await.unwrap().text, "frontier");
+    // A tiny request routes to cheap.
+    let small = LlmRequest {
+        messages: vec![RichMessage::Text {
+            role: "user".into(),
+            content: "x".into(),
+        }],
+        temperature: None,
+        max_tokens: None,
+        tools: vec![],
+        ..Default::default()
+    };
+    assert_eq!(fb.generate(small).await.unwrap().text, "cheap");
+}
+
+#[tokio::test]
+async fn cascade_escalates_structural_fail_under_background_smart() {
+    use mur_common::agent::AgentProfile;
+    use mur_common::config::{ModelSwitchConfig, SmartConfig};
+    // cheap returns InvalidResponse (structural), then primary succeeds.
+    let mut scripts = std::collections::HashMap::new();
+    scripts.insert(
+        "cheap".to_string(),
+        vec![Err(LlmError::InvalidResponse("empty".into()))],
+    );
+    scripts.insert("primary".to_string(), vec![Ok(())]);
+
+    let mut cfg = ModelSwitchConfig::default();
+    cfg.default = Some("primary".into());
+    cfg.smart = SmartConfig {
+        enabled: true,
+        cheap: Some("cheap".into()),
+        max_escalations: 1,
+    };
+    let fb = FallbackLlmClient::new_routed(
+        AgentProfile::default_for_tests(),
+        cfg,
+        factory_for(scripts),
+        retry0(),
+    );
+
+    let mut bg = LlmRequest::default();
+    bg.intent = RequestIntent::Background(BackgroundKind::Scheduled);
+    assert_eq!(fb.generate(bg).await.unwrap().text, "primary");
+}
+
+#[tokio::test]
+async fn interactive_invalid_response_stays_fatal() {
+    let mut scripts = std::collections::HashMap::new();
+    scripts.insert(
+        "a".to_string(),
+        vec![Err(LlmError::InvalidResponse("x".into()))],
+    );
+    scripts.insert("b".to_string(), vec![Ok(())]);
+    let fb = FallbackLlmClient::new(vec!["a".into(), "b".into()], factory_for(scripts), retry0());
+    // Interactive: InvalidResponse is Fatal → returns the error, never tries b.
+    let err = fb.generate(LlmRequest::default()).await.unwrap_err();
+    assert!(matches!(err, LlmError::InvalidResponse(_)));
+}
+
+#[tokio::test]
+async fn cascade_respects_max_escalations() {
+    use mur_common::agent::AgentProfile;
+    use mur_common::config::{ModelSwitchConfig, SmartConfig};
+    // Both cheap AND primary return InvalidResponse; max_escalations=1
+    // means only one escalation is allowed, so the second structural
+    // failure (on primary) must surface instead of looping past the cap.
+    let mut scripts = std::collections::HashMap::new();
+    scripts.insert(
+        "cheap".to_string(),
+        vec![Err(LlmError::InvalidResponse("empty".into()))],
+    );
+    scripts.insert(
+        "primary".to_string(),
+        vec![Err(LlmError::InvalidResponse("still empty".into()))],
+    );
+
+    let mut cfg = ModelSwitchConfig::default();
+    cfg.default = Some("primary".into());
+    cfg.smart = SmartConfig {
+        enabled: true,
+        cheap: Some("cheap".into()),
+        max_escalations: 1,
+    };
+    let fb = FallbackLlmClient::new_routed(
+        AgentProfile::default_for_tests(),
+        cfg,
+        factory_for(scripts),
+        retry0(),
+    );
+
+    let mut bg = LlmRequest::default();
+    bg.intent = RequestIntent::Background(BackgroundKind::Scheduled);
+    let err = fb.generate(bg).await.unwrap_err();
+    assert!(matches!(err, LlmError::InvalidResponse(_)));
+}
+
+#[tokio::test]
+async fn generate_without_telemetry_never_panics() {
+    // Every fixture above builds via `new`/`new_routed` with no
+    // `with_telemetry` call — `telemetry` stays `None`, so this just
+    // re-confirms the no-sink path is inert (no panic, normal result).
+    let mut s = HashMap::new();
+    s.insert("a".into(), vec![Ok(())]);
+    let fb = FallbackLlmClient::new(vec!["a".into()], factory_for(s), retry0());
+    let resp = fb.generate(LlmRequest::default()).await.unwrap();
+    assert_eq!(resp.text, "a");
+}
+
+#[tokio::test]
+async fn routed_generate_emits_routing_event_on_escalation() {
+    use mur_common::agent::AgentProfile;
+    use mur_common::config::{ModelSwitchConfig, SmartConfig};
+    // cheap structurally fails once, escalates to primary which succeeds
+    // — exercises the "escalated" outcome + attempts/escalations counts
+    // end-to-end through the real with_telemetry wiring.
+    let mut scripts = HashMap::new();
+    scripts.insert(
+        "cheap".to_string(),
+        vec![Err(LlmError::InvalidResponse("empty".into()))],
+    );
+    scripts.insert("primary".to_string(), vec![Ok(())]);
+
+    let mut cfg = ModelSwitchConfig::default();
+    cfg.default = Some("primary".into());
+    cfg.smart = SmartConfig {
+        enabled: true,
+        cheap: Some("cheap".into()),
+        max_escalations: 1,
+    };
+    let (tx, mut rx) = tokio::sync::mpsc::channel(4);
+    let fb = FallbackLlmClient::new_routed(
+        AgentProfile::default_for_tests(),
+        cfg,
+        factory_for(scripts),
+        retry0(),
+    )
+    .with_telemetry(tx, "coach");
+
+    let mut bg = LlmRequest::default();
+    bg.intent = RequestIntent::Background(BackgroundKind::Scheduled);
+    bg.messages = vec![RichMessage::Text {
+        role: "user".into(),
+        content: "do the thing".into(),
+    }];
+    let resp = fb.generate(bg).await.unwrap();
+    assert_eq!(resp.text, "primary");
+
+    let ev = rx.try_recv().expect("routing event emitted");
+    match ev {
+        Event::Routing {
+            agent,
+            intent,
+            model_ref,
+            reason,
+            outcome,
+            attempts,
+            escalations,
+            task_summary,
+            ..
+        } => {
+            assert_eq!(agent, "coach");
+            assert_eq!(intent, "background/scheduled");
+            assert_eq!(model_ref, "primary");
+            assert_eq!(reason, "smart-background");
+            assert_eq!(outcome, "escalated");
+            assert_eq!(attempts, 2);
+            assert_eq!(escalations, 1);
+            assert_eq!(task_summary, "do the thing");
+        }
+        other => panic!("expected Event::Routing, got {other:?}"),
+    }
+}
